moving old animate-in-out tests back
