--- src/view/use-animate-in-out/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { default } from './use-animate-in-out';
